+++ packages/isbot/index.js
@@ -1,11 +1,12 @@
 const isbot = require('isbot');
+const defaultProps = ['length', 'name', 'arguments', 'caller', 'prototype'];
 
 module.exports = function crawlerRequest(request, response, next) {
 	let is;
 
 	function get() {
 		return is = typeof is === 'boolean'
-			?	is
+			? is
 			: isbot(request.get('user-agent'))
 		;
 	}
@@ -15,4 +16,11 @@ module.exports = function crawlerRequest(request, response, next) {
 	next();
 };
 
-Object.assign(module.exports, isbot);
+Object.assign(
+	module.exports,
+	...Object.getOwnPropertyNames(isbot).filter(
+		prop => !defaultProps.includes(prop)
+	).map(
+		prop => ({ [prop]: isbot[prop] })
+	)
+);