+++ packages/isbot/spec.js
@@ -27,7 +27,7 @@ describe('isbot', async() => {
 		expect(request.isbot).to.be.true;
 	});
 	it('Should memoise the result if isbot from request', async() => {
-		request.headers['user-agent'] = 'Chrome';
+		request.headers['user-agent'] = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/94.0.4606.71 Safari/537.36';
 
 		expect(request.isbot).to.be.false;
 		request.headers['user-agent'] = 'Googlebot';