+++ packages/time/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 2.0.3
+## 2021-10-13
+Default to "*" if route is not defined
+
+# 2.0.2
+## 2020-10-13
+Support route set as array
+
 # 2.0.1
 ## 2019-08-25
 Convert Bigint to number only after calculation